refactor: 기물 배치 패키지 분리
